requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.27.0